# static/openapi.yaml
openapi: 3.0.0
info:
  title: EU Transit Declaration Generator API
  description: This is a simple message generator for the European project NCTS (New Computerized Transit System) Phase 5.
  version: 1.0.0
servers:
  - url: https://ncts5.monnier.cloud
  - url: http://localhost:8080
  
tags:
- name : CC007C messages
  description : arrival advice from the trader
- name : CC044C messages
  description : unloading remarks from the trader
- name : CC014C messages
  description: invalidation request from a holder of transit procedure

paths:
  
  /CC007C:
    get:
      tags:
      - CC007C messages
      summary: Retrieve a standard CC007C XML message in Austrian format
      parameters:
      - name: traderRIN
        in: query
        required: true
        schema:
          type: string
      - name: MRN
        in: query
        required: true
        schema:
          type: string
      - name: authorisationID_C522
        in: query
        required: true
        schema:
          type: string
      - name: OoDes
        in: query
        required: true
        schema:
          type: string
      - name: traderEORI
        in: query
        required: true
        schema:
          type: string
      - name: locationID
        in: query
        required: true
        schema:
          type: string          
      responses:
        "200":
          description: Successful response
          content:
            application/xml:
              schema:
                type: string

  /CC007C_TIR:
    get:
      tags:
      - CC007C messages
      summary: Retrieve a TIR CC007C XML message in Austrian format
      parameters:
      - name: traderRIN
        in: query
        required: true
        schema:
          type: string
      - name: MRN
        in: query
        required: true
        schema:
          type: string
      - name: authorisationID_C520
        in: query
        required: true
        schema:
          type: string
      - name: authorisationID_C522
        in: query
        required: true
        schema:
          type: string
      - name: OoDes
        in: query
        required: true
        schema:
          type: string
      - name: traderEORI
        in: query
        required: true
        schema:
          type: string
      - name: locationID
        in: query
        required: true
        schema:
          type: string          
      responses:
        "200":
          description: Successful response
          content:
            application/xml:
              schema:
                type: string

  /CC044C:
    get:
      tags:
      - CC044C messages
      summary: Retrieve a standard CC044C XML message in Austrian format
      parameters:
        - name: traderRIN
          in: query
          required: true
          schema:
            type: string 
        - name: MRN
          in: query
          required: true
          schema:
            type: string
        - name: OoDes
          in: query
          required: true
          schema:
            type: string
        - name: traderEORI
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Successful response
          content:
            application/xml:
              schema:
                type: string

  
  /CC014C:
    get:
      tags:
      - CC014C messages
      summary: Retrieve a standard CC014C XML message in Austrian format
      parameters:
        - name: holderRIN
          in: query
          required: true
          schema:
            type: string        
        - name: MRN
          in: query
          required: true
          schema:
            type: string
        - name: OoDep
          in: query
          required: true
          schema:
            type: string
        - name: holderEORI
          in: query
          required: true
          schema:
            type: string          
      responses:
        "200":
          description: Successful response
          content:
            application/xml:
              schema:
                type: string